Allow custom icons
