--- CardinalKit-Example/CardinalKit/Supporting Files/Assets.xcassets/CardinalKit/EmailIcon.imageset/Contents.json
@@ -1,12 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "email_icon.pdf",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
